requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-auth<2.42.0
pypdf
python-docx